requirements.txt
langchain
langchain-groq
faster-whisper
sounddevice
numpy
python-dotenv
pyttsx3